refactor: rename threeCameraService -> threeCamera.service
ref #2318
