+++ visualization/app/codeCharta/codeCharta.component.spec.ts
@@ -7,7 +7,7 @@ import { getService, instantiateModule } from "../../mocks/ng.mockhelper"
 import { InjectorService } from "./state/injector.service"
 import { StoreService } from "./state/store.service"
 import { setAppSettings } from "./state/store/appSettings/appSettings.actions"
-import { ThreeCameraService } from "./ui/codeMap/threeViewer/threeCameraService"
+import { ThreeCameraService } from "./ui/codeMap/threeViewer/threeCamera.service"
 import sample1 from "./assets/sample1.cc.json"
 import sample2 from "./assets/sample2.cc.json"
 import { CCFile, LayoutAlgorithm } from "./codeCharta.model"
