+++ visualization/app/codeCharta/ui/codeMap/codeMap.label.service.spec.ts
@@ -15,7 +15,7 @@ import {
 	SpriteMaterial,
 	Vector3
 } from "three"
-import { ThreeCameraService } from "./threeViewer/threeCameraService"
+import { ThreeCameraService } from "./threeViewer/threeCamera.service"
 import { ThreeSceneService } from "./threeViewer/threeSceneService"
 import { getService, instantiateModule } from "../../../../mocks/ng.mockhelper"
 import { StoreService } from "../../state/store.service"
