+++ visualization/app/codeCharta/ui/codeMap/codeMap.mouseEvent.service.spec.ts
@@ -2,7 +2,7 @@ import "./codeMap.module"
 import "../../codeCharta.module"
 import { IRootScopeService, IWindowService } from "angular"
 import { ClickType, CodeMapMouseEventService, CursorType } from "./codeMap.mouseEvent.service"
-import { ThreeCameraService } from "./threeViewer/threeCameraService"
+import { ThreeCameraService } from "./threeViewer/threeCamera.service"
 import { ThreeSceneService } from "./threeViewer/threeSceneService"
 import { getService, instantiateModule } from "../../../../mocks/ng.mockhelper"
 import { ThreeRendererService } from "./threeViewer/threeRenderer.service"
