+++ visualization/app/codeCharta/ui/codeMap/threeViewer/threeViewerService.ts
@@ -1,5 +1,5 @@
 import { ThreeSceneService } from "./threeSceneService"
-import { ThreeCameraService } from "./threeCameraService"
+import { ThreeCameraService } from "./threeCamera.service"
 import { ThreeOrbitControlsService } from "./threeOrbitControlsService"
 import { ThreeRendererService } from "./threeRenderer.service"
 import { ThreeStatsService } from "./threeStats.service"
