+++ visualization/app/codeCharta/ui/ribbonBar/showScenariosButton/scenario.service.spec.ts
@@ -2,7 +2,7 @@ import { MatDialog } from "@angular/material/dialog"
 import { State } from "../../../state/angular-redux/state"
 import { Store } from "../../../state/angular-redux/store"
 import { setState } from "../../../state/store/state.actions"
-import { ThreeCameraService } from "../../codeMap/threeViewer/threeCameraService"
+import { ThreeCameraService } from "../../codeMap/threeViewer/threeCamera.service"
 import { ThreeOrbitControlsService } from "../../codeMap/threeViewer/threeOrbitControlsService"
 import { ErrorDialogComponent } from "../../dialogs/errorDialog/errorDialog.component"
 import { ScenarioService } from "./scenario.service"
