+++ visualization/app/codeCharta/ui/viewCube/viewCube.component.spec.ts
@@ -2,7 +2,7 @@ import "./viewCube.module"
 import { ThreeOrbitControlsService } from "../codeMap/threeViewer/threeOrbitControlsService"
 import { ViewCubeMouseEventsService } from "./viewCube.mouseEvents.service"
 import { PerspectiveCamera } from "three/src/cameras/PerspectiveCamera"
-import { ThreeCameraService } from "../codeMap/threeViewer/threeCameraService"
+import { ThreeCameraService } from "../codeMap/threeViewer/threeCamera.service"
 import { Color, Mesh, Vector3, WebGLRenderer } from "three"
 import { OrbitControls } from "three/examples/jsm/controls/OrbitControls"
 import { ViewCubeComponent } from "./viewCube.component"
